GitHub Packager Support
* Change stats command to be a PassCommand
* Move CommonOptions methods to CommonOptionsHelper
* Forward looking changes to upcoming sbt-ossuminc
