+++ commands/src/main/scala/com/ossuminc/riddl/commands/FromCommand.scala
@@ -5,6 +5,7 @@
  */
 
 package com.ossuminc.riddl.commands
+
 import com.ossuminc.riddl.language.CommonOptions
 import com.ossuminc.riddl.language.Messages.Messages
 import com.ossuminc.riddl.passes.PassesResult
@@ -65,7 +66,7 @@ class FromCommand extends CommandPlugin[FromCommand.Options](FromCommand.cmdName
     outputDirOverride: Option[Path]
   ): Either[Messages, PassesResult] = {
     val loadedCO =
-      CommandOptions.loadCommonOptions(options.inputFile.fold(Path.of(""))(identity)) match {
+      CommonOptionsHelper.loadCommonOptions(options.inputFile.fold(Path.of(""))(identity)) match {
         case Right(newCO: CommonOptions) =>
           if commonOptions.verbose then {
             println(
