+++ project/plugins.sbt
@@ -1,4 +1,4 @@
-addSbtPlugin("com.ossuminc" % "sbt-ossuminc" % "0.8.3")
+addSbtPlugin("com.ossuminc" % "sbt-ossuminc" % "0.9.2-0-868ced6e-20240326-1318")
 
 // This enables sbt-bloop to create bloop config files for Metals editors
 // Uncomment locally if you use metals, otherwise don't slow down other
