+++ stats/src/main/scala/com/ossuminc/riddl/stats/StatsCommand.scala
@@ -6,41 +6,63 @@
 
 package com.ossuminc.riddl.stats
 
-import com.ossuminc.riddl.commands.InputFileCommandPlugin
+import com.ossuminc.riddl.commands.{CommandOptions, InputFileCommandPlugin, PassCommand, PassCommandOptions}
 import com.ossuminc.riddl.language.Messages.Messages
 import com.ossuminc.riddl.language.CommonOptions
-import com.ossuminc.riddl.passes.{Pass, PassesResult, Riddl}
+import com.ossuminc.riddl.passes.Pass.standardPasses
+import com.ossuminc.riddl.passes.{Pass, PassInput, PassesCreator, PassesOutput, PassesResult, Riddl}
 import com.ossuminc.riddl.utils.Logger
+import scopt.OParser
+import pureconfig.{ConfigCursor, ConfigReader}
 
+import java.io.File
 import java.nio.file.Path
 
-/** Validate Command */
-class StatsCommand extends InputFileCommandPlugin("stats") {
-  import InputFileCommandPlugin.Options
+object StatsCommand {
+  val cmdName: String = "stats"
+  case class Options(
+    inputFile: Option[Path] = None
+  ) extends PassCommandOptions
+      with CommandOptions {
+    def command: String = cmdName
+    def outputDir: Option[Path] = None
+  }
+}
 
-  override def run(
-    options: Options,
-    commonOptions: CommonOptions,
-    log: Logger,
-    outputDirOverride: Option[Path]
-  ): Either[Messages, PassesResult] = {
-    options.withInputFile { (inputFile: Path) =>
-      val passes = Pass.standardPasses ++ Seq({ (input, output) => StatsPass(input, output) })
-      Riddl.parseAndValidatePath(inputFile, commonOptions, passes = passes, logger = log) match {
-        case Left(messages) => Left(messages)
-        case Right(result) =>
-          result.outputOf[StatsOutput](StatsPass.name) match {
-            case Some(stats) =>
-              println(s"Maximum Depth: ${stats.maximum_depth}")
-              for (k, v) <- stats.categories do {
-                println(s"$k: $v")
-              }
-              println()
-            case None => println("No statistics generated")
-          }
-          Right(result)
-      }
+/** Stats Command */
+class StatsCommand extends PassCommand[StatsCommand.Options]("stats") {
+  import StatsCommand.Options
+
+  // Members declared in com.ossuminc.riddl.commands.CommandPlugin
+  def getConfigReader: ConfigReader[Options] = { (cur: ConfigCursor) =>
+    for
+      topCur <- cur.asObjectCursor
+      topRes <- topCur.atKey(pluginName)
+      objCur <- topRes.asObjectCursor
+      inFileRes <- objCur.atKey("input-file").map(_.asString)
+      inFile <- inFileRes
+    yield {
+      Options(inputFile = Some(Path.of(inFile)))
     }
+  }
+
+  def getOptions: (OParser[Unit, Options], com.ossuminc.riddl.stats.StatsCommand.Options) = {
+    import builder.*
+    cmd(StatsCommand.cmdName)
+      .children(
+        arg[File]("input-file")
+          .action { (file, opt) =>
+            opt.copy(inputFile = Some(file.toPath))
+          }
+          .text("The main input file on which to generate statistics.")
+      )
+      .text("Loads a configuration file and executes the command in it") ->
+      StatsCommand.Options()
+  }
+
+  // Members declared in com.ossuminc.riddl.commands.PassCommand
+  def overrideOptions(options: Options, newOutputDir: Path): Options = {
+    options // we don't support overriding the output dir
   }
 
   override def replaceInputFile(
@@ -56,4 +78,9 @@ class StatsCommand extends InputFileCommandPlugin("stats") {
       resolveInputFileToConfigFile(options, commonOptions, configFile)
     }
   }
+
+  override def getPasses(log: Logger, commonOptions: CommonOptions, options: Options): PassesCreator = {
+    standardPasses :+ StatsPass.creator
+  }
+
 }
